Renames and add kind
