+++ .github/workflows/build.yaml
@@ -5,7 +5,7 @@ on:
     branches: [ main ]
   pull_request:
     branches: [ main ]
-  
+
 env:
   PROJECT_NAME: Boxer.Validator.Nginx
   REGISTRY: ghcr.io
@@ -21,13 +21,13 @@ jobs:
       contents: read       # required for dependabot PRs
     steps:
       - uses: actions/checkout@v4
-        
+
       - name: Install minimal stable with clippy and rustfmt
         uses: actions-rust-lang/setup-rust-toolchain@v1.13.0
         with:
-         toolchain: stable
-         components: rustfmt, clippy
-         
+          toolchain: stable
+          components: rustfmt, clippy
+
       - name: Install just, cargo-llvm-cov, cargo-nextest
         uses: taiki-e/install-action@v2.56.13
         with:
@@ -35,10 +35,21 @@ jobs:
 
       - name: Rustfmt Check
         uses: actions-rust-lang/rustfmt@v1
-        
+
+      - name: Create k8s Kind Cluster
+        uses: helm/kind-action@v1
+        with:
+          cluster_name: 'kind'
+
+      - name: Check kind cluster
+        run: |
+          kubectl cluster-info
+          kubectl get nodes
+          kind get kubeconfig --name kind
+
       - name: Generate code coverage
         run: cargo llvm-cov --all-features --workspace --lcov --output-path lcov.info
-        
+
       - name: Upload coverage to Codecov
         uses: romeovs/lcov-reporter-action@v0.4.0
         with:
